Reduce language choice buttons Also show actual language in button, not alternative language
Related to OAFWM-52
