+++ Resources/Private/Scss/_base.scss
@@ -85,3 +85,11 @@ ul {
   list-style-type: none;
   padding: 0;
 }
+
+.-shown {
+  display: flex !important;
+}
+
+.-hidden {
+  display: none !important;
+}
