+++ Resources/Private/Templates/Partials/Header.html
@@ -14,6 +14,7 @@
 		</a>
 	</div>
 	<div class="head__links">
+
 		<ul class="head__list">
 			<li>
 				<f:link.page pageUid="{START_UID}" title="{f:translate(key:'startpage', extensionName: 'tmpl_ipoa')}"
@@ -93,8 +94,8 @@
 							</svg>
 							<span class="hidden-sub-l">
 								<img
-									src='<f:uri.image src="EXT:tmpl_ipoa/Resources/Public/Images/flags/{ipoa:getAlternativeLanguageFlag()}.png" />'>
-										<ipoa:getAlternativeLanguage/>
+									src='<f:uri.image src="EXT:tmpl_ipoa/Resources/Public/Images/flags/{ipoa:getLanguageFlag()}.png" />'>
+										<ipoa:getLanguage/>
 							</span>
 						</a>
 					</li>
