+++ Resources/Private/Templates/Start.html
@@ -154,8 +154,8 @@
 									</svg>
 									<span class="hidden-sub-l">
 										<img
-											src='<f:uri.image src="EXT:tmpl_ipoa/Resources/Public/Images/flags/{ipoa:getAlternativeLanguageFlag()}.png" />'>
-											<ipoa:getAlternativeLanguage/>
+											src='<f:uri.image src="EXT:tmpl_ipoa/Resources/Public/Images/flags/{ipoa:getLanguageFlag()}.png" />'>
+											<ipoa:getLanguage/>
 									</span>
 								</a>
 							</li>
@@ -168,7 +168,7 @@
 						<v:page.languageMenu
 							defaultLanguageLabel="<ipoa:getDefaultLanguage />"
 							linkCurrent="0"
-							defaultIsoFlag="dede"
+							defaultIsoFlag="de"
 							classCurrent="'start__link--active'"
 							hideNotTranslated="{true}">
 							<div class="start__language-links js-start-language-links">
